requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.7.4
python-docx==1.1.2
pymupdf==1.24.2
pytesseract==0.3.10
Pillow==10.3.0
unstructured==0.14.7
sentence-transformers==3.0.1
chromadb==0.5.5
pandas==2.2.2
openpyxl==3.1.2
weasyprint==62.3
requests==2.32.3